Flags moved to separate packages
