+++ pkg/cmd/gobarista/flags/flags.go
@@ -1,5 +1,5 @@
 // Copyright 2023 PraserX
-package main
+package flags
 
 import "github.com/urfave/cli/v2"
 
@@ -8,4 +8,10 @@ var FlagConfig = cli.StringFlag{
 	Value:   ".local/config.ini",
 	Usage:   "path to configuration file",
 	Aliases: []string{"c"},
+}
+
+var FlagPrettyPrint = cli.BoolFlag{
+	Name:    "pretty",
+	Aliases: []string{"p"},
+	Usage:   "print formatted output",
 }
