+++ pkg/cmd/gobarista/main.go
@@ -5,6 +5,7 @@ import (
 	"os"
 
 	"github.com/praserx/gobarista/pkg/cmd/gobarista/commands"
+	"github.com/praserx/gobarista/pkg/cmd/gobarista/flags"
 	"github.com/praserx/gobarista/pkg/logger"
 	"github.com/praserx/gobarista/pkg/version"
 
@@ -22,7 +23,7 @@ func main() {
 		Description: "This is a simple coffee billing application - for enterprise, non-profits or government.",
 		Copyright:   "(c) Praser",
 		Flags: []cli.Flag{
-			&FlagConfig,
+			&flags.FlagConfig,
 		},
 		Commands: []*cli.Command{
 			&commands.Billing,
